requirements.txt
# ============================================================================
# Python Version Compatibility
# ============================================================================
# This requirements.txt is tested with Python 3.8, 3.9, 3.10, and 3.11
# Python 3.12+ may have compatibility issues with TensorFlow 2.15.0
# Recommended: Python 3.10 or 3.11

# ============================================================================
# Core Web Framework (Required for Flask application)
# ============================================================================
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# ============================================================================
# Deep Learning Framework (CRITICAL - Required for loading .h5 model files)
# ============================================================================
# TensorFlow 2.15.0 includes Keras 2.15.0 internally
# These versions are specifically tested for compatibility with saved .h5 models
tensorflow==2.15.0

# ============================================================================
# Numerical Computing (Compatible with TensorFlow 2.15.0)
# ============================================================================
# NumPy < 2.0 required for TensorFlow 2.15.0
numpy==1.24.3
scipy==1.11.4

# ============================================================================
# Image Processing (Required for PIL/Pillow image operations)
# ============================================================================
Pillow==10.1.0

# ============================================================================
# Data Processing (Required for pandas DataFrame operations)
# ============================================================================
pandas==2.1.3

# ============================================================================
# Visualization (Required for Plotly charts)
# ============================================================================
plotly==5.18.0
kaleido==0.2.1

# ============================================================================
# Model File Support (CRITICAL - Required for loading .h5 model files)
# ============================================================================
h5py==3.10.0

# ============================================================================
# TensorFlow Core Dependencies (Auto-installed with TensorFlow, but pinned for consistency)
# ============================================================================
protobuf==4.25.1
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
flatbuffers==23.5.26
gast==0.5.4
google-pasta==0.2.0
grpcio==1.60.0
idna==3.6
libclang==16.0.6
opt-einsum==3.3.0
packaging==23.2
requests==2.31.0
setuptools==69.0.2
six==1.16.0
termcolor==2.4.0
typing-extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0


